[codeshuffle] - moved xbmc.cpp/.h to platform
