+++ xbmc/platform/Makefile.in
@@ -1,6 +1,7 @@
 .SUFFIXES : .m .mm .cpp
 
 SRCS = posix/MessagePrinter.cpp
+SRCS += xbmc.cpp
 SRCS += XbmcContext.cpp
 
 ifeq ($(findstring osx,@ARCH@), osx)
