+++ xbmc/platform/xbmc.cpp
