+++ xbmc/platform/xbmc.h
